Flatten all the way through, and simplify a helper (#42)
This is a quick little PR that picks up a loose thread and simplifies a
helper.

1. We flattened the AST and massaged away ADTs that only had one
constructor, but this also meant that we needed to stop expecting those
constructors and just expect the payloads instead. Maybe that had
already been done but had gone away during a merge.
2. I found a way to simplify a piece of OCaml but I realized this a
little too late and the relevant PR (#37) had already closed!