Update Interpreter To Include New Policies (#44)
#37 gave us an interpreter and well-formedness checker. #41 gave us the
ability to express non-work conserving policies in the DSL.

Unfortunately, the parallel progress on both of these tracks means that
the interpreter, along with pretty-printing functionality, was released
without accounting for the latest policies in the DSL. This PR attempts
to fix that by extending the functionality of the interpreter. This also
introduces adds test scripts for the programs for the new policies,
cleans up the code for evaluation and pretty printing and ensures that
the types are properly flattened throughout the evaluator.