Compiler: convert policy to JSON (#75)
This PR makes progress towards the compiler project whose aim is to
generate Calyx code from Rio programs. In json.ml, the program takes a
Rio program and returns the JSON representation of it using the yojson
library. A test suite can be found in `tests/serialization`.

Moving forward, I think it makes sense for the rest of the
implementation for the compiler to be in the calyx repo. One thing that
I was unsure about is to 'link' the two together, as in, how does the
JSON get generated so that it can be passed to the python calyx eDSL?
Not sure if this is more something for down the road.